+++ src/main/java/lk/ijse/food_ordering_backend/entity/OrderItem.java
@@ -5,8 +5,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import java.io.Serializable;
-import java.math.BigDecimal;
+ 
 
+// Entity class representing an order item in the food ordering system.
 @Entity
 @Table(name = "order_item")
 @Data
@@ -31,5 +32,5 @@ public class OrderItem implements Serializable {
     private Integer quantity;
 
     @Column(nullable = false)
-    private BigDecimal price;
+    private Double price;
 }
